requirements.txt
# Core LangChain dependencies
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langsmith>=0.1.0

# Data processing and validation
pydantic>=2.0.0
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Web frameworks
fastapi>=0.104.0
uvicorn>=0.24.0

# Streamlit for UI (optional)
streamlit>=1.28.0

# Additional utilities
typing-extensions>=4.7.0